+++ mods/BrewLAN/documentation/Changelog.md
@@ -605,6 +605,10 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
   * Fixed an Iyadesu exploit where you could reclaim the drone and keep the recipes forever.
   * Fixed an exploit with the lattice shield where you could reclaim the tower and keep the drones.
 
+  ### 2021-02-15
+
+  * R&D: Fixed an AI not being able to upgrade research centres.
+
 ## 2019-10-13|0.8.9
 
 * Incremented numbers for static public release.